backend/requirements.txt
alembic==1.7.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.1
cryptography==3.4.8
environs==9.3.4
Flask==2.0.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.4.0
Flask-JWT-Extended==4.2.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.13.0
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.2.0
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.25
urllib3==1.26.7
Werkzeug==2.0.2
